+++ lib/cli.dart
@@ -1,2 +1,3 @@
-@Deprecated("Use import 'package:stepflow/io.dart'; instead.")
+@Deprecated("Use 'package:stepflow/io.dart' instead. "
+    "'cli.dart' will be removed in the next major version.")
 export 'io.dart';