+++ lib/common.dart
@@ -1,2 +1,3 @@
-@Deprecated("Use import 'package:stepflow/core.dart'; instead.")
+@Deprecated("Use 'package:stepflow/core.dart' instead. "
+    "'common.dart' will be removed in the next major version.")
 export 'core.dart';