+++ lib/platform.dart
@@ -1,3 +1,6 @@
+@Deprecated("stepflow's platform has benn moved over to its dart package "
+    "stepflow_platform. Will be removed from stepflow in the next major version."
+    "The parts of the library will be removed the next major version.")
 export 'src/platform/attributes/android.dart';
 export 'src/platform/attributes/baremetal.dart';
 export 'src/platform/attributes/ios.dart';